Face Extraction/requirements.txt
tkinter
custom-tkinter
opencv-python==4.8.0
pillow==10.2.0
facenet-pytorch==2.6.0
pyinstaller==6.9.0
